auth/requirements.txt
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
Flask==3.0.2
Flask-Cors==4.0.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jwcrypto==1.5.6
MarkupSafe==2.1.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymongo==4.6.3
pyparsing==3.1.2
requests==2.31.0
rsa==4.9
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
